refactor: Extract shared KSP utilities and improve SchemaExtensionProcessor readability
- Moved `filterByRootPackage` and `getSchemaParameters` to a new `kspUtils` file for reuse across KSP components.
- Simplified `SchemaExtensionProcessor` by delegating parameter retrieval and root package filtering to utilities.
- Added `schemaAnnotationDefaults` for consistent default parameter handling.
- Enabled parallel test execution for `SchemaExtensionProcessor` tests by marking them as `@Isolated`.
