+++ kotlinx-schema-generator-core/src/test/kotlin/kotlinx/schema/generator/reflect/ReflectionIntrospectorTest.kt
@@ -29,6 +29,7 @@ class ReflectionIntrospectorTest {
     )
 
     @Description("Available colors")
+    @Suppress("unused")
     enum class Color { RED, GREEN, BLUE }
 
     data class WithEnum(
